+++ full-moon/src/ast/mod.rs
@@ -10,6 +10,12 @@ use full_moon_derive::{Node, Visit};
 use lua52::*;
 #[cfg(feature = "lua54")]
 use lua54::*;
+
+#[cfg(any(feature = "luau", feature = "cfxlua"))]
+mod compound;
+#[cfg(any(feature = "luau", feature = "cfxlua"))]
+pub use compound::*;
+
 #[cfg(feature = "luau")]
 use luau::*;
 pub use parser_structs::AstResult;
@@ -39,7 +45,7 @@ mod versions;
 
 #[cfg(any(feature = "lua52", feature = "luajit"))]
 pub mod lua52;
-#[cfg(feature = "lua54")]
+#[cfg(any(feature = "lua54", feature = "cfxlua"))]
 pub mod lua54;
 /// A block of statements, such as in if/do/etc block
 #[derive(Clone, Debug, Default, Display, PartialEq, Node, Visit)]
@@ -204,6 +210,16 @@ pub enum Field {
         equal: TokenReference,
         /// The `value` part of `name = value`
         value: Expression,
+    },
+
+    /// A set constructor field, such as .a inside { .a } which is equivalent to { a = true }
+    #[display("{dot}{name}")]
+    #[cfg(feature = "cfxlua")]
+    SetConstructorField {
+        /// The `.` part of `.a`
+        dot: TokenReference,
+        /// The `a` part of `.a`
+        name: TokenReference,
     },
 
     /// A field with no key, just a value (such as `"a"` in `{ "a" }`)
@@ -391,7 +407,7 @@ pub enum Stmt {
 
     /// A compound assignment, such as `+=`
     /// Only available when the "luau" feature flag is enabled
-    #[cfg(feature = "luau")]
+    #[cfg(any(feature = "luau", feature = "cfxlua"))]
     #[display("{_0}")]
     CompoundAssignment(CompoundAssignment),
     /// An exported type declaration, such as `export type Meters = number`
@@ -413,11 +429,11 @@ pub enum Stmt {
 
     /// A goto statement, such as `goto label`
     /// Only available when the "lua52" or "luajit" feature flag is enabled.
-    #[cfg(any(feature = "lua52", feature = "luajit"))]
+    #[cfg(any(feature = "lua52", feature = "luajit", feature = "cfxlua"))]
     Goto(Goto),
     /// A label, such as `::label::`
     /// Only available when the "lua52" or "luajit" feature flag is enabled.
-    #[cfg(any(feature = "lua52", feature = "luajit"))]
+    #[cfg(any(feature = "lua52", feature = "luajit", feature = "cfxlua"))]
     Label(Label),
 }
 
@@ -1695,7 +1711,7 @@ impl LocalAssignment {
             #[cfg(feature = "luau")]
             type_specifiers: Vec::new(),
             name_list,
-            #[cfg(feature = "lua54")]
+            #[cfg(any(feature = "lua54", feature = "cfxlua"))]
             attributes: Vec::new(),
             equal_token: None,
             expr_list: Punctuated::new(),
@@ -1737,7 +1753,7 @@ impl LocalAssignment {
     /// `local foo <const>, bar, baz <close>` returns an iterator containing:
     /// `Some(Attribute("const")), None, Some(Attribute("close"))`
     /// Only available when the "lua54" feature flag is enabled.
-    #[cfg(feature = "lua54")]
+    #[cfg(any(feature = "lua54", feature = "cfxlua"))]
     pub fn attributes(&self) -> impl Iterator<Item = Option<&Attribute>> {
         self.attributes.iter().map(Option::as_ref)
     }
@@ -1760,7 +1776,7 @@ impl LocalAssignment {
     }
 
     /// Returns a new LocalAssignment with the given attributes
-    #[cfg(feature = "lua54")]
+    #[cfg(any(feature = "lua54", feature = "cfxlua"))]
     pub fn with_attributes(self, attributes: Vec<Option<Attribute>>) -> Self {
         Self { attributes, ..self }
     }
@@ -1786,7 +1802,7 @@ impl LocalAssignment {
 
 impl fmt::Display for LocalAssignment {
     fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        #[cfg(feature = "lua54")]
+        #[cfg(any(feature = "lua54", feature = "cfxlua"))]
         let attributes = self.attributes().chain(std::iter::repeat(None));
         #[cfg(not(feature = "lua54"))]
         let attributes = std::iter::repeat_with(|| None::<TokenReference>);