+++ full-moon/src/ast/parsers.rs
@@ -8,6 +8,12 @@ use super::{
     span::ContainedSpan,
     Expression, FunctionBody, Parameter,
 };
+
+#[cfg(any(feature = "cfxlua", feature = "luau"))]
+use super::{
+    Var
+};
+
 use crate::{
     ast,
     node::Node,
@@ -40,6 +46,7 @@ pub fn parse_block(state: &mut ParserState) -> ParserResult<ast::Block> {
                 let semicolon = state.consume_if(Symbol::Semicolon);
                 stmts.push((stmt, semicolon));
             }
+
             ParserResult::Value(StmtVariant::LastStmt(last_stmt)) => {
                 let semicolon = state.consume_if(Symbol::Semicolon);
                 let last_stmt = Some((last_stmt, semicolon));
@@ -112,6 +119,25 @@ enum StmtVariant {
     #[allow(unused)]
     LastStmt(ast::LastStmt),
 }
+
+#[cfg(any(feature = "luau", feature = "cfxlua"))]
+fn parse_compound_assignment(state: &mut ParserState, var: Var) -> ParserResult<StmtVariant> {
+    let compound_operator = state.consume().unwrap();
+
+    let ParserResult::Value(expr) = parse_expression(state) else {
+        state.token_error(compound_operator, "expected expression to set to");
+        return ParserResult::LexerMoved;
+    };
+
+    return ParserResult::Value(StmtVariant::Stmt(ast::Stmt::CompoundAssignment(
+        ast::CompoundAssignment {
+            lhs: var,
+            compound_operator: ast::CompoundOp::from_token(compound_operator),
+            rhs: expr,
+        },
+    )));
+}
+
 
 fn parse_stmt(state: &mut ParserState) -> ParserResult<StmtVariant> {
     let Ok(current_token) = state.current() else {
@@ -303,21 +329,24 @@ fn parse_stmt(state: &mut ParserState) -> ParserResult<StmtVariant> {
                             || token.is_symbol(Symbol::CaretEqual)
                             || token.is_symbol(Symbol::TwoDotsEqual)) =>
                 {
-                    let compound_operator = state.consume().unwrap();
-
-                    let ParserResult::Value(expr) = parse_expression(state) else {
-                        state.token_error(compound_operator, "expected expression to set to");
-                        return ParserResult::LexerMoved;
-                    };
-
-                    return ParserResult::Value(StmtVariant::Stmt(ast::Stmt::CompoundAssignment(
-                        ast::CompoundAssignment {
-                            lhs: var,
-                            compound_operator: ast::CompoundOp::from_token(compound_operator),
-                            rhs: expr,
-                        },
-                    )));
+                    return parse_compound_assignment(state, var);
                 }
+
+                #[cfg(feature = "cfxlua")]
+                Ok(token)
+                if state.lua_version().has_cfxlua()
+                    && (token.is_symbol(Symbol::PlusEqual)
+                    || token.is_symbol(Symbol::MinusEqual)
+                    || token.is_symbol(Symbol::StarEqual)
+                    || token.is_symbol(Symbol::SlashEqual)
+                    || token.is_symbol(Symbol::CaretEqual)
+                    || token.is_symbol(Symbol::LeftShift)
+                    || token.is_symbol(Symbol::RightShift)
+                    || token.is_symbol(Symbol::BitwiseAndAssignment)
+                    || token.is_symbol(Symbol::BitwiseOrAssignment)) =>
+                    {
+                        return parse_compound_assignment(state, var);
+                    }
 
                 Ok(token) if token.is_symbol(Symbol::Comma) || token.is_symbol(Symbol::Equal) => {}
 
@@ -518,7 +547,7 @@ fn parse_stmt(state: &mut ParserState) -> ParserResult<StmtVariant> {
             })))
         }
 
-        #[cfg(any(feature = "lua52", feature = "luajit"))]
+        #[cfg(any(feature = "lua52", feature = "luajit", feature = "cfxlua"))]
         TokenType::Symbol {
             symbol: Symbol::Goto,
         } => {
@@ -1011,13 +1040,13 @@ fn expect_local_assignment(
 
     #[cfg(feature = "luau")]
     let mut type_specifiers = Vec::new();
-    #[cfg(feature = "lua54")]
+    #[cfg(any(feature = "lua54", feature = "cfxlua"))]
     let mut attributes = Vec::new();
 
     for name in names.into_pairs() {
         let (name, punctuation) = name.into_tuple();
 
-        #[cfg(feature = "lua54")]
+        #[cfg(any(feature = "lua54", feature = "cfxlua"))]
         attributes.push(name.attribute);
 
         #[cfg(feature = "luau")]
@@ -1040,10 +1069,18 @@ fn expect_local_assignment(
         attributes,
     };
 
-    local_assignment.equal_token = match state.consume_if(Symbol::Equal) {
+    #[cfg(not(feature = "cfxlua"))]
+    let symbols = [Symbol::Equal];
+
+    #[cfg(feature = "cfxlua")]
+    let symbols = [Symbol::Equal, Symbol::In];
+
+
+    local_assignment.equal_token = match state.consume_if_symbols(&symbols) {
         Some(equal_token) => Some(equal_token),
         None => return Ok(local_assignment),
     };
+
 
     match parse_expression_list(state) {
         ParserResult::Value(expr_list) => local_assignment.expr_list = expr_list,
@@ -1190,6 +1227,27 @@ fn force_table_constructor(
                     equal: equal_token,
                     value,
                 }
+            }
+
+            // used for assignments like this { .a }, its equivalent to { a=true }
+            #[cfg(feature = "cfxlua")]
+            TokenType::Symbol {
+                symbol: Symbol::Dot,
+            } => {
+                let dot = state.consume().unwrap();
+
+                let key = match state.current() {
+                    Ok(token) if token.token_kind() == TokenKind::Identifier => state.consume().unwrap(),
+                    Ok(token) => {
+                        state.token_error(token.clone(), "expected identifier after `.`");
+                        return unfinished_table(left_brace, fields);
+                    }
+                    Err(()) => {
+                        return unfinished_table(left_brace, fields);
+                    }
+                };
+
+                ast::Field::SetConstructorField { dot, name: key }
             }
 
             _ => {
@@ -1504,6 +1562,27 @@ fn parse_suffix(state: &mut ParserState) -> ParserResult<ast::Suffix> {
     };
 
     match current.token_type() {
+        #[cfg(feature = "cfxlua")]
+        TokenType::Symbol {
+            symbol: Symbol::SafeNavigation,
+        } => {
+            let safe_navigation = state.consume().unwrap();
+            let name = match state.current() {
+                Ok(token) if token.token_kind() == TokenKind::Identifier => {
+                    state.consume().unwrap()
+                }
+
+                Ok(_) => {
+                    state.token_error(safe_navigation, "expected identifier after `?.`");
+                    return ParserResult::LexerMoved;
+                }
+
+                Err(()) => return ParserResult::LexerMoved,
+            };
+
+            ParserResult::Value(ast::Suffix::Index(ast::Index::Dot { dot: safe_navigation, name }))
+        }
+
         TokenType::Symbol {
             symbol: Symbol::Dot,
         } => {
@@ -3216,7 +3295,7 @@ fn expect_generic_type_params(
 #[derive(Clone)]
 struct Name {
     name: TokenReference,
-    #[cfg(feature = "lua54")]
+    #[cfg(any(feature = "lua54", feature = "cfxlua"))]
     attribute: Option<super::lua54::Attribute>,
     #[cfg(feature = "luau")]
     type_specifier: Option<ast::TypeSpecifier>,
@@ -3278,16 +3357,16 @@ fn force_name(_state: &mut ParserState, name: TokenReference) -> Name {
     }
 }
 
-#[cfg(feature = "lua54")]
+#[cfg(any(feature = "lua54", feature = "cfxlua"))]
 fn force_name_with_attributes(state: &mut ParserState, name: TokenReference) -> Name {
     // NOTE: whenever attributes can be parsed, type specifiers are possible
     // so we should fall back to parsing type specifiers if an attribute is not found.
     // NOTE: we do not attempt to parse both type specifiers and attributes at the same time
-    if !state.lua_version().has_lua54() {
+    if !state.lua_version().has_lua54() && !state.lua_version().has_cfxlua() {
         return force_name_with_type_specifiers(state, name);
     }
 
-    #[cfg(feature = "lua54")]
+    #[cfg(any(feature = "lua54", feature = "cfxlua"))]
     if let Some(left_angle_bracket) = state.consume_if(Symbol::LessThan) {
         const ERROR_INVALID_ATTRIBUTE: &str = "expected identifier after `<` for attribute";