+++ full-moon/tests/cfxlua_cases/fail/parser/bitwise_or_assignment_operator/source.lua
@@ -1,0 +1,2 @@
+local num = 4
+1 |= num